requirements.txt
# Génération de données
faker==24.11.0

# Manipulation de données
pandas==2.2.0
numpy==1.26.0

# Bases de données
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
# presto-python-client==0.8.4
trino
# HDFS
hdfs==2.7.0
pyarrow==15.0.0

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Utilitaires
colorama==0.4.6
tabulate==0.9.0
python-dateutil==2.8.2

# Tests
pytest==7.4.3